cleanup for first "public" version
This commit adds the usual suspects for "public" version: Rakefile
with a :setup target, embryo of documentation. Moreover, the use
of exceptions is a bit cleaned up a bit. Now, Orocos::NotFound is
thrown every time we try to get a new CORBA reference and the object
does not exist.