test: restructure and move to minitest
The minitest library is now going to be standard in both
Ruby 1.9 and Rubinius. We therefore move the test suite
to there (given how small it was anyway). Moreover, the
modules are not automatically built anymore but instead
are built through the rake "setup_tests" task, to avoid
very long test times.