roby: use model tags in the hierarchy
Until now, we were using the hack that a given data source model
should be associated to one concreate task model (not a tag), and
this model would be used in the plan.

Instead, we now use the abstract model tag to represent the
dependencies, and develop the data source model a bit further
to support that. Namely, a data source can be associated with
an "interface" model, which represents what interface this data
source has. Concrete tasks that implement the source need to have
this interface implemented, but do not have to be subclasses of
the orogen model.