roby: detect need for port mapping during instanciation
We may have to do port mapping when the user explicitely
selects a task, as the task may not have the desired data
source interface as a main interface.

Properly detect that case, and for now raise NotImplementedError.

This requires to change the selection mechanism so that it is
made as late as possible (it was done early before).