fix broken task order in Process.spawn
During the integration of valgrind, the set of names were converted
into a Set, so that we can merge the names from valgrind. It obviously
broke the list of processes that Process.spawn yields.

Remove the convertion to set, and stop auto-adding the processes
that are listed in the :valgrind option, as knowing the order
of process object would be unpractical (i.e. getting the corresponding
Process object reliably would be next to impossible)