roby: implement port mapping at instanciation time
If the user selects a specific instance using #use, and using
this instance requires to do port mapping, this is now properly
done.

Example:
  - one stereocamera that can also output both images
  - one stereoprocessing component that processes two images and
    outputs a point cloud
  - a Stereo composition: (StereoProcessing camera0 camera1)

Then, if the cameras exported by StereoCamera are selected for
the Stereo composition, we need to map the port names from the
generic "Camera" model into the specific port names of StereoCamera
(for instance imageLeft and imageRight).

The convention is to use the name of the slave data sources. Assuming
that a camera data source has one "image" output port, the following
declaration:

  class StereoCamera
    driver_for "stereo"
    data_source "camera", :as => 'left', :slave_of => "stereo"
  end

will make the system search for a port named "leftImage", "stereoLeftImage",
"imageLeft" and "imageStereoLeft".