roby: handle the merge of already instanciated devices
That was one of the goals of the last two weeks of work:
to handle one task that handle multiple devices at the same
time, and to recognize that if that task is instanciated
two times, then they are mergeable if the device names are
not assigned yet (or are identical).

For the occasion, #can_replace? is replaced by #can_merge?.
While a.can_replace?(b) says if a can replace b without any
modification, #can_merge? acknowledges that a must be modified
before it can take the place of b, and that this modification
can be done without breaking other dependencies.