roby: cleanup the name/object mess for model handling
until now, the user API was mostly designed around the use
of names. It is very error prone, and led to having two
registries for models: the actual constants (i.e. the
Orocos::RobyPlugin::* namespaces) and the SystemModel
attributes.

Remove the SystemModel attribute and assume that everything
is stored in the namespaces (global model definitions).
Moreover, in most cases, it is assumed that the user will
give us the model object. It is planned to use sandboxed
environments with const_missing() to make it easier for the
user.