roby: use find_local_tasks instead of find_tasks in instanciation code
First, this allows to pick to what the algorithms apply by adding
the relevant tasks in the transaction, instead of having
them apply it on everything.

Moreover, it is important as tasks that have been merged
are now removed from the transaction, and therefore from
the scope of the merge algorithms.