make Port.validate_policy(Port.validate_policy(policy)) work
validate_policy always sets a size of 0, so that the C code
does not have to handle a missing :size field. However, it was
also checking that there was no :size field for data policies.

Thus, calling validate_policy repeatedly on a given connection
policy was raising. This is obviously wrong and is fixed by this
commit.