roby: rework the dynamic connection handling
The main change is that changing connections is now a transaction-
oriented operation: it is either all changes or no changes.

To achieve that, the new scheme uses three graphs. The task
relation DataFlows is the user interface, that allows to
specify the local routes between ports (i.e. both connections
and forwardings).

The RequiredDataFlow *graph* is a translation of DataFlows in which
only concrete connections are registered.

Finally, ActualDataFlow is also a *graph* in which the connections
that actually exist between the components are registered.