roby: we must call #specialize in #constrain
The reason why is that, by using #constrain, the user actually says
"well, I want to use this interface at that place". This leads to
an implicit expectation that the connections will be done "right".

The needed change is that multiple calls to #specialize now work:
the block will be simply applied on subsequent calls. Moreover,
#constrain gains an ':exclusive' option. If set, the :not option
of #specialize will be set accordingly