roby: be more robust during disconnection
First of all, it is not required to get an existing CORBA port
for disconnection. Having the CORBA proxy is enough, so use
only the corba proxy without checking that the port exists.

Moreover, try both disconnection methods. I.e. disconnect from
the output first, and then do it on the input if the output cannot
be contacted.