roby: TaskContext#stop may fail if the task transitioned to fatal_error first
This commit has two parts. The first one calls #reset_error in the event command
to make sure that we *can* start the task. This is only a workaround though.

The current hypothesis I have, w.r.t. this bug, is that interrupt is called at
the same time than the task transitions to fatal_error. Therefore, #stop fails,
and the task goes into internal_error.

This commit rescues the StateTransitionFailed, and if the task *is* in fatal
error, simply lets the poll block do the rest. Otherwise, we do re-raise the
exception.