roby: compute port mappings at declaration time
Until now, port mappings were computed on-demand. Since we
want to be able to do more complex mappings, it is not
sustainable anymore.

Compute port mappings in #data_service, and save them
in the new ProvidedDataService structure. Also, add
automatic port mappings if the ports in the underlying
component does not match the ports in the service
definition.