roby: be more permissive when solving loops in the merge graph
If we have a two-task loop, we now arbitrarily remove one if they are equivalent
from the p.o.v. of merge_sort_order. This is a *really* stupid policy, but is
valid nonetheless. Its only drawback would be that some merge situations that could
be solved may not be. It is still better than what we have now (rejecting the
situation right away)