roby: try to be a bit smarter w.r.t. cycle removal
We know start by iteratively applying simple merges and merge_prepare
to remove cycles where there is a transitive relationship (i.e.

  a <=> b <=> c

where the relationship of b and c is well defined, but a and b
are equivalent. The algorithm will not first do b => c, then
reconsider a <=> c and successfully merge.

Then, we do a simple random cycle breaking, in the event that it
would be enough.

Worst case: the resolution will fail at the task allocation stage,
as we won't have enough deployments.