roby: fix transitionning between components
The main issue was that we were not clearing the relations
of already used tasks. This had the effect of refusing their
merge at the end of #resolve.

We now have a clear two-stage process, where in the first stage
the plan represents what is needed, which is then applied to the
deployed tasks.

In-between the two stages, tasks that are in the plan and not in
the transaction are "cleared", i.e. any I/O dataflow and dependencies
are removed to allow the merge to work.