roby: we must allow a child to be overriden, even at the same level of the model hierarchy
This check has already been removed, but has been reintroduced
during the recent code cleanup. Add a comment in the code to
explain the reason (reproduced below)

We do NOT check for an already existing definition. The reason
is that specialization (among other) will add a default child,
that may be overriden by the composition's owner. Either to
set arguments, or to have a specialization over an aspect of a
data service use a more specific task model in the specialized
composition.

Anyway, the remainder checks that the new definition is a
valid overloading of the previous one.