std::string has a special place in operations
The opportunistic loading of typekits relies on callin
the getModel operation, which itself returns a string.

This is a problem, as /std/string is not defined as long
as we did not load any orogen-generated typekit. So, to
bootstrap the process, we actually need to use the RTT
typekit, in which a string is 'string'

So, convert /std/string to string in operation constructor