roby: introduce a method to create private specializations of task models
The idea is that, to represent dynamic port instanciation better,
one should be able to create a "singleton task model" that will
be used for single tasks.

One can then add new ports to this singleton model without changing
the main model it is part of.

This led to the ability to do

  device(MainDevice).
     slave(SlaveDevice)

where the SlaveDevice definition leads to creation of new ports on the
driver for MainDevice.

This is meant to replace part of the combus implementation in the
(hopefully near) future.