roby: streamline DataSource/Service/ComBus model structure and configuration API
First things first. One main problem with the DataSource/DataService/
ComBus models is that, in order to get the model-level methods on the
submodels, we had to use ModuleExtension. Which was leaking to the
final classes and other model types (in the end). A much better
solution is actually to create specialized DataSourceModel and
ComBusDriverModel subclasses of DataServiceModel for those, and make
sure that a submodel of type X is always an instance of X or one of
its subclass.

Second of all, the #device_id and #set API to configure devices sucks.
Big time. Offer a way for the data source models and com bus models to
propertly extend that API so that the device configuration looks more
natural (and is more checked "at the source")