roby: move the runtime connection management out of Engine
This commit also fixes a bug introduced by b3c440421bb374f48eecadb84e39b14706655701

Namely, the system was not checking if all static inputs of a task
were connected before running it. It is now reintroduced, and
Engine#all_inputs_connected? is move to the definition of DataFlow,
where it belongs