mitigate the effects of the CORBA asynchronicity in RTT2 w.r.t. state transitions
The problem is that it can be confusing in Ruby scripts that just
after

  task.start

task.state is not :RUNNING. It makes writing Ruby scripts that look
for state very hard (not a problem in Roby though, as it handles this
case nicely).

To mitigate that effect, the state transition calls now wait, by default,
for a state transition to happen before returning.