Revert "roby: fix missing device detection for tasks that drive multiple devices"
The rest of the execution layer actually assumes that, when a task
drives multiple devices, then it should run only if all said devices
are mapped onto a device.

Reestablish this check, as otherwise we might generate non-executable
plans. The proper fix would be to find a way to tell the supervision
to 'ignore' devices, but I am not even sure that it is actually needed
in practice.

This reverts commit 06c5476426840245c9ad7fa0dbe0276dc6d61bb9.