dont't convert the C++ object to ruby on reads that did not read anything
in #read_new, if no new sample is available, there is no need to convert the
C++ type back to ruby.

Moreover, converting it is harmful in cases where the C++ type is an
opaque (namely, for smart pointers), as the C++ sample can then be
empty and the opaque-to-typelib convertion functions will see a NULL
pointer.