roby: make the usage of "instance specifications" more pervasive and less confusing
This commit changes the behaviour of the #use method, and makes it
available at quite a few different points as well.

On the one hand, #use() statements are not recursive anymore. It means that
when someone does

  add(Cmp::Bla).
    use(Task)

then Task will not be passed as a valid selection to the compositions
included in Bla. The main reason is that it gets quite tricky to avoid
recursions (where a selection in use() is passed to a composition that
is being instanciated because of the use() statement. Ouch.

On the other hand, a few mechanisms have been added to make this new
state usable.

First of all, one can call #use in compositions that are listed in #use
statements, as for instance:

  add(Cmp::CorridorServoing).
    use(Cmp::Odometry.use('imu'))

Second of all, defines and #use in compositions are available when specifying
global selections, as in

  use Cmp::Odometry => Cmp::Odometry.use('imu')

Third of all, a composition can forward a selection to its children,
forcing the child to use the same component than the parent when it
makes sense.

  composition 'CorridorServoing' do
    add Srv::Pose, :as => 'pose'
    add(Cmp::ControlLoop).use(pose)
  end

Note that this selection can still be overriden at instanciation
time by providing suitable #use information