roby: always add the root composition to the set of to-be-specialized compositions
Otherwise, a specialization A will *not* be recursively applied on specializations
created after it, i.e. in

  compo.specialize A
  compo.specialize B

only A+B would be registered. Then, if one does

  compo.specialize C

If B+A cannot be specialized with C, C will not be 'mixed in' A and B alone
(i.e. A+C or B+C won't be created)