roby: if the activity is not periodic, assume that we are triggered at the device's rates
This commit fixes a bug for components that are triggered internally.

It basically assumes that a component that is not periodic *and* is
associated with devices will trigger itself at the 'right' period.

Obviously, this assumption might be wrong in general. In these cases,
people will have to use the burst parameters to fine-tune the dataflow
computation for their task. However, for the majority of tasks, the
assumption holds.

Signed-off-by: Sylvain Joyeux <sylvain.joyeux@dfki.de>