roby: fix creation of placeholder tasks
The main issue that this commit fixes was that placeholder tasks
did not have their own name. In the end, model.name was returning
DataServiceModel, which was breaking the dRoby code (distributed
and logging marshalling / demarshalling)

This commit factors out the creation of placeholder tasks for
composition children and data services. It also makes the creation
of placeholder task streamlined by handling every case (including
the case where models == [task_model]