roby: when updating the task states, stop as soon as the state is STOPPED
The issue here is that in some (very) bad cases, runtime states
are being sent after STOPPED, which leads to some issues since we
start emitting events on stopped tasks.

So, filter those out for the time being as they should not make roby
stop.

It happened in a component that used a separate thread, where the
thread was not properly synchronized with stopHook