roby: make sure #reusable? only applies to the already configured tasks
The issue is that #reusable? was checking on
#needs_configuration?, which is an attribute that applies
to all Roby tasks deployed on the same Orocos task.

The net effect was that the merge algorithm got confused, as
new Roby tasks deployed on a certain orocos task were marked
as not reusable.

Fix that by checking setup? as well