roby: do a static garbage collection phase at the end of generate_system_network
Otherwise, "named" tasks -- such as devices -- are still present in the
plan *and* taken into account during the merging process. This can lead
to issues when one uses an alternate configuration for a device for
instance, as the device might get merged into the associated deployed
task, leading to the actual task having no deployment at all.