roby: make the selection tasks compatible with replacement
This fixes an issue where, if the toplevel task of a selected
instance gets replaced (as e.g. due to a redeployment), the old
toplevel task would remove the instance requirement from the
engine, but the new task would still be in the plan with a
mission flag.

It would lead to a completely inconsistent state in the engine and
a few problems (obviously)