roby: fix find_most_specialized to resolve parent-children in the possible candidates
Right now, find_most_specialized_compositions was returning all
compositions that could be selected *given the explicit child
selection*.

However, it could lead to returning [a, b} where b is a
specialization of a. Change the policy to return a if b cannot
be selected by the composition. If a and b are unrelated, still
return both.