roby: major cleanup of instance requirement vs. instanciation handling
This commit moves all the handling of instance requirement and
selection representation to separate classes, making the handling
of composition child in line with the "normal" instance spec. This
will allow to simply extend InstanceRequirement to extend the
capabilities of the system from a modeling point of view.

It also removes quite a few special cases that were here in Composition
just to deal with the schizophrenic instance spec / composition children
situation.

Moreover, the dependency selection / management is now made more explicit
by the introduction of the DependencyInjection* classes. It allows a more
explicit control over the selection priorities, and in the end will allow
for "selection contexts" (as: "for these tasks, use by default everything
related to the left arm)