make validate_policy *really* a validation function
Until now, it was adding default values. It is meant to only validate
a policy, not set it to some default value.

This is important in the roby plugin, as otherwise the update_merge_info
method will fill in empty policies with the default policy, even though
an empty policy has a meaning during plan building.