fix in replay with property log files
the patch for overriding of log files didn't work, if the properties.log
was loaded before other streams. This is because the named task is already
registered in this case, but the file_path not set (since it comes from
a property). checking if the file_path is set in the task fixes the problem.