replay: don't use Orocos.registry by default
It was a workaround of a long-time solved limitation. Before,
values had to be coming from types of the same registry to be
replayed on ports. This is not true anymore: they only have
to be of compatible types.

This solves some issues where the replay loading mechanism
would fail because of incompatible types, even though the
types were only used in streams that were not replayed.