roby: fix weird behaviour in Engine#add(model.use(...))
This problem stems from the issue that we don't differentiate
requirements and default choices in compositions. Therefore, when
one does

  add(Cmp::Odometry).
    use(Srv::Orientation => 'imu')

in a composition, the first part (Cmp::Odometry) is taken as a
requirement, and is used for dependency injection, while the second
part is not. InstanceRequirements#selection_for therefore accesses
the #models method on criteria that are InstanceRequirements.

This is OK so far from a usability point of view (even though it is
probably annoying in complex cases), but no in the engine. Indeed,

  use Cmp::Odometry => Cmp::Odometry.use(Odometry::FirstTask)
  add(Cmp::Odometry.use(Odometry::SecondTask))

would add the general odometry, not the specified second one.

This commit fixes the issue by resolving user selection in Engine#add
only if the passed criteria is a plain model or name