fix dependency specification in composition instantiation
We were using the plain Hash#merge to merge instance-level and
model-level dependency specifications. This is wrong as, once
normalized, the role set (for instance) should be merged instead
of being replaced.

Use the facilities provided by DependencyGraphClass to merge
the various dependency specifications. The only part that gets
overriden is the success and failure event sets, as we really
want to override and not merge in this case