ros: improve behaviour w.r.t. the asynchronous nature of the ROS graph update
The issue covered here is that we read the ROS graph asynchronously.
There is therefore a delay between an action (say, the creation
of a topic) and when #find_output_port will return the value.

This commit tries to reduce the side effects by adding explicit
synchronization points wherever an object is queried and not found.

It means that #each_input_port and #each_output_port are e.g. still
affected, though. Unfortunately, reading the ROS graph is too slow
to be done every time.