be compatible with having base types handled by an orogen project
So far, we had hardcoded the fact that the typekit included in RTT
itself was providing the base types. Since it is barely the case,
and has dire consequences ('array' being 'vector</double>' anyone ?),
we introduced the base/orogen/std orogen project.

This commit makes sure that we use the info from base/orogen/std when
available.