TaskContextProxy: add method_missing and respond_to_missing
If the underlying task is reachable it is now possible to call methods
only defined on the task object. In the case the task is not reachable
respond_to will return false.

All calls are called synchronously. Therefore the call blocks until
no other thread is using the underlying bject