fix RubyDeployment
ProcessBase stores the set of tasks for this particular deployment.
This broke RubyDeployment since RubyDeployment was already using
the tasks attribute as the set of tasks that this deployment contains.

To fit with the process-servers-should-have-their-own-nameservice
scheme, we now add a Local name service to RubyProcessServer, and
use it in RubyDeployment#task to resolve the tasks using the
base class' #task method.