use basename to generate the logger file name
If one sets the name server IP to 127.0.0.1 explicitly, the task names start
having a prefix. This was causing the log file names to get a prefix of e.g.
127.0.0.1/ as well and them not being created at all (since the directory did
not exist)

For the time being, use basename to generate the log file name, which gets us
back to the old behaviour.