make Orocos.clear only reset the loaders instead of creating new ones
Since Orocos.default_loader is really used as a singleton, it gets
hard when a new object is created (since that information has to
be propagated to all "users" of it). In particular, in syskit, it
makes being robust to reload pretty hard.

Instead, clear all caches in all loaders when .clear is called. As
a side-effect, in order to make #clear a true reset, we have to
define a separate DefaultLoader class which sets up the default
loader properly (i.e. adds the RTT definitions and does the type exports
on #clear).