fix Orocos.initialized? not being reset after calling Orocos.clear
Orocos.initialized? was returning the value of Orocos::CORBA.initialized?
However, the latter will never be reset (did not find a way to do that
reliably), while the former should be reset when Orocos.clear is called,
as we do reset some of the objects (most notably, we dispose of
Orocos.ruby_task).