add sanity check to verify that one does not try starting the same process twice
This makes sure that we won't overwrite an already existing process
object by a new one with the same name. It does *not* double up with the
check in Orocos::Process#spawn, as it would trigger in the following
scenario:

 - start process X
 - call kill for process X
 - start process X while we have not received the death announce from
   the server (yet) for the previous instance of X

In this scenario, the current code would announce the death of the *new*
instance (since it has replaced the old instance in the processes hash)
and Orocos::Process#spawn would not raise as we already killed the old
instance.

The new code will rightly raise when #start is called