test: cleanup common test setup and migrate to minitest
This cleans up the common test setup to a more "modern" layout
(read: following the rest of rock's common test setup). It removes
a lot of cruft from the existing test files (good thing) and migrates
to minitest. The minitest migration is due to the minitest gem being
incompatible with the version shipped by Ruby itself, meaning
that test::unit is basically unusable. Since minitest *is* the default
test framework, simply use it directly