remote_processes: fix loader handling in Client
While it is important to verify that the remote process has a
deployment / task model when required, the code must use the
root loader to load things (and NOT the specific loader). Otherwise,
we might end up with multiple versions of the same model because
they got loaded from different sources.