TaskConfiguration: do not save metadata inside of config-hashes
We added a early version of metadata in rock, metadata are represented as
propertys. Therefore the configurations include this. But the metadata
should not be handled throught the normal configuration handling.
Therefore ignoring them here.

This solves the issue that oroconc extracts the metadata information of
tasks.

This is a interrim solution, see: https://github.com/rock-core/tools-orocosrb/issues/16