remove usage of Dir.chdir
Dir.chdir is really to be avoided as much as possible. In ruby 1.9+,
all (most?) subprocess-spawning methods such as system accept an
argument to tell where the subprocess should be executed. As it
is done in the fork (and not in the parent), it is guaranteed to
not affect the parent in any way -- something that chdir never did.