log: only decode data for streams that are listening to it
This modifies Replay#step and the Log::* classes to lazy-read
the samples from the files instead of reading them eagerly. In
addition, the infrastructure got modified to properly propagate
raw samples and convert to ruby values on-demand.

This really optimizes the common case of processing only a few
streams very well.

I've only taken care of the incremental case (i.e. add new listeners
dynamically). The decremental case (where listeners are removed) is
as always a lot more complex and currently really not on the hot path.