Add ability to change log level per deployment
Orocos.run was extended with the :log_level option. This allows
to set the log level of base/logging from the script. The parameter
can either be one log level for all deployments (e.g. :INFO), or can
be a Hash, to control the level for each deployment.

:log_level => { 'control' => :DEBUG }

e.g. will set the log level for the control deployment to DEBUG.