ruby_tasks: move stubbing of operations to a separate StubTaskContext class
While properties and ports are really RTT objects, operations are
currently not implemented as such (i.e. cannot be called remotely for
instance). There is however an interest in faking them for testing.

There was some very limited stubbing of operations in
TaskContext#setup_from_orogen_model. Move that out of there in order
to not give the (wrong) idea that operations are supported by real
ruby tasks. Create a StubTaskContext class that stubs operations more
completely (implementing the #operation method and the Operation API).
Finally, modify Process and ProcessManager to allow creating
StubTaskContext instances instead of TaskContext.