remove rescue of Orocos::CORBAError in has_port?
This was introduced as the (wrong) bugfix of a deeper nature on MacOSX,
where do_has_port? would raise CORBAError instead of NotFound

Even though it *will* break existing libraries, this has to disappear.
It is currently completely misrepresenting dead tasks by saying that
they do not have a port instead of reporting that they are simply
*not there*. I noticed it because it had bad at runtime in Syskit.