ruby_tasks: improve interaction of StubTaskContext with flexmock
The problem this is aiming to solve is to give room for the stubbed
task user to override methods. The current way was putting methods
at or near the highest priority. Defining them in a module allows to
put them below the singleton class, and thus allowing flexmock to override
them by prepending to the singleton class.