add #resolve_all_tasks to resolve a process tasks
The current method, #wait_running is really designed for the script use
case, but is actually problematic to build upon. Moreover, it requires
that the tasks be resolved twice: one during #wait_running and one by
the caller.

In itself, it's not a problem in scripts, but it's fairly bad if the
remote accesses need to be controlled in more advanced systems.

A new API, #resolve_all_tasks, is parallel to the existing one. It
either returns the resolved tasks if the process is ready, or nil if
it is not. It can save partial resolutions in a provided map to
avoid re-resolving tasks that are known to be accessible.