log: don't filter out ports that have no samples
If the port is used, it should be present on the aligner. The
refactored pocolog started (rightly) raising if the stream
aligner is being queried about a stream it's not aligning,
and this broke Vizkit.

Basically, vizkit items are enumerating all ports. Because
pocolog master is simply returning zero and nil for streams
that are not being aligned, Vizkit was happily showing empty
streams and the fact that we're not validating stream aligner's
input was silently ignored.

# Conflicts:
#	test/log/test_replay.rb