improve control over Process#kill
The current implementation of Process#kill does two things behind
the scenes:

- stop and/or cleanup the tasks
- send a signal (which might be SIGINT or SIGKILL or anything really)

In a first iteration, I mis-used the fact that the implementation
would bypass the cleanup stage if a signal is explicitely given. This
is obviously (in hindsight) very specific to Orocos::Process and neither
applies to ruby tasks or unmanaged tasks (in Syskit).

This commit modified the first iteration to instead pass functional
arguments, `cleanup` and `hard` which give finer-grained control over
the process.