fix: workaround multi-threading issue related to typekit loading
So ...

Turns out the typekit code is simply not MT-safe. We're essentially
and probably saved by Ruby's global lock, but issues do pop up.

This commit works around a problem that is now prominent with the
Syskit IDE caused by recent changes that streamlines type loading
(and trigger typekits not being loaded while they were). This
is only a quick workaround, more work is coming