feat: read IOR of a task context from pipe
By reading the IOR of a task context from a pipe, we are not limited by
CORBA name service to resolve the tasks. The wait running method must be
called to register the IORs for each task in the process. This is done
by reading the message from the pipe, validating it and saving it as a
Hash. Later, this Hash is used to resolve the tasks when the
resolve_all_tasks method is called. Since this method might be called
before wait_running is ran, a nonblocking wait_running is called in it to
ensure that the IORs are registered.

Overall, the expected pipeline for Syskit is:

The process client calls the process server, which spawns the Orocos::Process.
On spawn, the IOR pipe is created. Then, the execution engine triggers the
deployment ready resolution, which calls for the process' wait_running,
expecting the IOR mappings to be resolved. They are registered and then
resolve_all_tasks is called. If the IORs are valid, the deployment resolve
the remote task handles and hopefully is becomes ready.