fix: handle a race condition when accessing CORBA on dying remote processes
If the call happens before the connection is actually closed, OmniORB
handles it as a timeout. Otherwise, it is a ComError.

This is actually not-so-great in our case, as we really do not want
the timeout to happen. Note that it does not apply to crashing
components, so the general Syskit behavior should not be affected.