fix: partially reintroduce sanity check on OutputReader's read path
This partially reverts commit c7f516d451dc99c951f6d631913dedb32b527982.

cf252a453d52a480a8dee4f94b6948fe111bdda7 changed the name of the "core"
read method. This method was overloaded in OutputReader to raise if
the remote process was dead, and disconnect the read side.

I think the boat has sailed on {#read} not raising, and I'm not
willing to re-introduce that behavior. However, I do believe
the disconnect_all is a good move: it cleans up, and makes sure
that a caller that cares does get notified that the reader is
dead.