fix: support partial models
Dynamic ports and/or the process' own ruby task context are both
good examples of a reason why we should support it.

But, unlike orocos.rb, do optimize if a model is available by internally
relying on the existence of a model. We create said model at the
object creation boundary using remote info.